Move tests to their own file and check results
